@@ -25,22 +25,23 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "illuminate/support": "~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0|^8.0",
-        "guzzlehttp/guzzle": "~5.0|~6.0|^7.0"
+        "php": ">=7.3",
+        "illuminate/support": "^8.0|^9.0|^10.0|^11.0",
+        "guzzlehttp/guzzle": "^7.0",
+        "laravel/pint": "dev-main"
     },
     "autoload": {
         "psr-4": {
-            "Softon\\Indipay\\": "src/"
+            "Abounaja\\Indipay\\": "src/"
         }
     },
     "extra": {
         "laravel": {
             "providers": [
-                "Softon\\Indipay\\IndipayServiceProvider"
+                "Abounaja\\Indipay\\IndipayServiceProvider"
             ],
             "aliases": {
-                "Indipay": "Softon\\Indipay\\Facades\\Indipay"
+                "Indipay": "Abounaja\\Indipay\\Facades\\Indipay"
             }
         }
     },
