@@ -1,9 +1,13 @@
-<?php namespace Softon\Indipay\Facades;
+<?php
+
+namespace Abounaja\Indipay\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
-class Indipay extends Facade {
-
-    protected static function getFacadeAccessor() { return 'indipay'; }
-
+class Indipay extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return 'indipay';
+    }
 }
