@@ -1,24 +1,34 @@
-<?php namespace Softon\Indipay\Gateways;
+<?php
 
+namespace Abounaja\Indipay\Gateways;
+
+use Abounaja\Indipay\Exceptions\IndipayParametersMissingException;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\View;
-use Softon\Indipay\Exceptions\IndipayParametersMissingException;
 
-class CCAvenueGateway implements PaymentGatewayInterface {
+class CCAvenueGateway implements PaymentGatewayInterface
+{
+    protected $parameters = [];
 
-    protected $parameters = array();
     protected $merchantData = '';
+
     protected $encRequest = '';
+
     protected $testMode = false;
+
     protected $workingKey = '';
+
     protected $accessCode = '';
+
     protected $liveEndPoint = 'https://secure.ccavenue.com/transaction/transaction.do?command=initiateTransaction';
+
     protected $testEndPoint = 'https://test.ccavenue.com/transaction/transaction.do?command=initiateTransaction';
+
     public $response = '';
 
-    function __construct()
+    public function __construct()
     {
         $this->workingKey = Config::get('indipay.ccavenue.workingKey');
         $this->accessCode = Config::get('indipay.ccavenue.accessCode');
@@ -32,20 +42,20 @@ class CCAvenueGateway implements PaymentGatewayInterface {
 
     public function getEndPoint()
     {
-        return $this->testMode?$this->testEndPoint:$this->liveEndPoint;
+        return $this->testMode ? $this->testEndPoint : $this->liveEndPoint;
     }
 
     public function request($parameters)
     {
-        $this->parameters = array_merge($this->parameters,$parameters);
+        $this->parameters = array_merge($this->parameters, $parameters);
 
         $this->checkParameters($this->parameters);
 
-        foreach($this->parameters as $key=>$value) {
+        foreach ($this->parameters as $key => $value) {
             $this->merchantData .= $key.'='.$value.'&';
         }
 
-        $this->encRequest = $this->encrypt($this->merchantData,$this->workingKey);
+        $this->encRequest = $this->encrypt($this->merchantData, $this->workingKey);
 
         return $this;
 
@@ -58,31 +68,29 @@ class CCAvenueGateway implements PaymentGatewayInterface {
     {
 
         Log::info('Indipay Payment Request Initiated: ');
-        return View::make('indipay::ccavenue')->with('encRequest',$this->encRequest)
-                             ->with('accessCode',$this->accessCode)
-                             ->with('endPoint',$this->getEndPoint());
+
+        return View::make('indipay::ccavenue')->with('encRequest', $this->encRequest)
+            ->with('accessCode', $this->accessCode)
+            ->with('endPoint', $this->getEndPoint());
 
     }
 
-
     /**
      * Check Response
-     * @param $request
+     *
      * @return array
      */
     public function response($request)
     {
         $encResponse = $request->encResp;
 
-        $rcvdString = $this->decrypt($encResponse,$this->workingKey);
+        $rcvdString = $this->decrypt($encResponse, $this->workingKey);
         parse_str($rcvdString, $decResponse);
 
         return $decResponse;
     }
 
-
     /**
-     * @param $parameters
      * @throws IndipayParametersMissingException
      */
     public function checkParameters($parameters)
@@ -102,81 +110,70 @@ class CCAvenueGateway implements PaymentGatewayInterface {
         }
 
     }
-    
-     /**
+
+    /**
      * CCAvenue encryption
-     * @param $plainText string
-     * @param $key string
+     *
+     * @param  $plainText  string
+     * @param  $key  string
      * @return string
      */
     public function encrypt($plainText, $key)
     {
         $key = $this->hextobin(md5($key));
-        $initVector = pack("C*", 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f);
+        $initVector = pack('C*', 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F);
         $openMode = openssl_encrypt($plainText, 'AES-128-CBC', $key, OPENSSL_RAW_DATA, $initVector);
         $encryptedText = bin2hex($openMode);
+
         return $encryptedText;
     }
-    
 
     /**
      * CCAvenue decryption
-     * @param $encryptedText string
-     * @param $key
+     *
+     * @param  $encryptedText  string
      * @return string
      */
     public function decrypt($encryptedText, $key)
     {
         $key = $this->hextobin(md5($key));
-        $initVector = pack("C*", 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f);
+        $initVector = pack('C*', 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F);
         $encryptedText = $this->hextobin($encryptedText);
         $decryptedText = openssl_decrypt($encryptedText, 'AES-128-CBC', $key, OPENSSL_RAW_DATA, $initVector);
+
         return $decryptedText;
     }
-  
-
 
     /**
-     * @param $plainText
-     * @param $blockSize
      * @return string
      */
     protected function pkcs5_pad($plainText, $blockSize)
     {
         $pad = $blockSize - (strlen($plainText) % $blockSize);
-        return $plainText . str_repeat(chr($pad), $pad);
+
+        return $plainText.str_repeat(chr($pad), $pad);
     }
 
-
     /**
-     * @param $hexString
      * @return string
      */
     protected function hextobin($hexString)
     {
         $length = strlen($hexString);
-        $binString="";
-        $count=0;
-        while($count<$length)
-        {
-            $subString =substr($hexString,$count,2);
-            $packedString = pack("H*",$subString);
-            if ($count==0)
-            {
-                $binString=$packedString;
+        $binString = '';
+        $count = 0;
+        while ($count < $length) {
+            $subString = substr($hexString, $count, 2);
+            $packedString = pack('H*', $subString);
+            if ($count == 0) {
+                $binString = $packedString;
+            } else {
+                $binString .= $packedString;
             }
 
-            else
-            {
-                $binString.=$packedString;
-            }
-
-            $count+=2;
+            $count += 2;
         }
+
         return $binString;
     }
-
-
-
-
 }
