@@ -1,25 +1,35 @@
-<?php namespace Softon\Indipay\Gateways;
+<?php
 
+namespace Abounaja\Indipay\Gateways;
+
+use Abounaja\Indipay\Exceptions\IndipayParametersMissingException;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\View;
-use Softon\Indipay\Exceptions\IndipayParametersMissingException;
 
-class InstaMojoGateway implements PaymentGatewayInterface {
+class InstaMojoGateway implements PaymentGatewayInterface
+{
+    protected $parameters = [];
 
-    protected $parameters = array();
     protected $merchantData = '';
+
     protected $encRequest = '';
+
     protected $testMode = false;
+
     protected $api_key = '';
+
     protected $auth_token = '';
+
     protected $liveEndPoint = 'https://www.instamojo.com/api/1.1/';
+
     protected $testEndPoint = 'https://test.instamojo.com/api/1.1/';
+
     public $response = '';
 
-    function __construct()
+    public function __construct()
     {
         $this->testMode = Config::get('indipay.testMode');
         $this->api_key = Config::get('indipay.instamojo.api_key');
@@ -28,30 +38,33 @@ class InstaMojoGateway implements PaymentGatewayInterface {
         $this->parameters['redirect_url'] = url(Config::get('indipay.instamojo.redirectUrl'));
     }
 
-    public function getEndPoint($param='')
+    public function getEndPoint($param = '')
     {
-        $endpoint = $this->testMode?$this->testEndPoint:$this->liveEndPoint;
+        $endpoint = $this->testMode ? $this->testEndPoint : $this->liveEndPoint;
+
         return $endpoint.$param;
     }
 
     public function request($parameters)
     {
-        $this->parameters = array_merge($this->parameters,$parameters);
+        $this->parameters = array_merge($this->parameters, $parameters);
 
         $this->checkParameters($this->parameters);
 
-        $client = new \GuzzleHttp\Client();
-        $response = $client->post($this->getEndPoint('payment-requests/'),
-                                        [
-                                            'headers'=> array(
-                                                'X-Api-Key' => $this->api_key,
-                                                'X-Auth-Token' => $this->auth_token,
-                                            ),
-                                            'form_params' => $this->parameters,
-                                        ])->getBody()->getContents();
+        $client = new \GuzzleHttp\Client;
+        $response = $client->post(
+            $this->getEndPoint('payment-requests/'),
+            [
+                'headers' => [
+                    'X-Api-Key' => $this->api_key,
+                    'X-Auth-Token' => $this->auth_token,
+                ],
+                'form_params' => $this->parameters,
+            ]
+        )->getBody()->getContents();
         $response = json_decode($response);
 
-        if($response->success){
+        if ($response->success) {
             $this->response = $response;
         }
 
@@ -66,15 +79,15 @@ class InstaMojoGateway implements PaymentGatewayInterface {
     {
 
         Log::info('Indipay Payment Request Initiated: ');
+
         //dd($this->response->payment_request->longurl);
-        return View::make('indipay::instamojo')->with('longurl',$this->response->payment_request->longurl);
+        return View::make('indipay::instamojo')->with('longurl', $this->response->payment_request->longurl);
 
     }
 
-
     /**
      * Check Response
-     * @param $request
+     *
      * @return array
      */
     public function response($request)
@@ -82,17 +95,19 @@ class InstaMojoGateway implements PaymentGatewayInterface {
         $payment_request_id = Request::input('payment_request_id');
         $payment_id = Request::input('payment_id');
 
-        $client = new \GuzzleHttp\Client();
-        $response = $client->get($this->getEndPoint('payment-requests/'.$payment_request_id.'/'.$payment_id.'/'),
+        $client = new \GuzzleHttp\Client;
+        $response = $client->get(
+            $this->getEndPoint('payment-requests/'.$payment_request_id.'/'.$payment_id.'/'),
             [
-                'headers'=> array(
+                'headers' => [
                     'X-Api-Key' => $this->api_key,
                     'X-Auth-Token' => $this->auth_token,
-                ),
-            ])->getBody()->getContents();
+                ],
+            ]
+        )->getBody()->getContents();
         $response = json_decode($response);
 
-        if($response->success){
+        if ($response->success) {
             return $response;
         }
 
@@ -100,9 +115,7 @@ class InstaMojoGateway implements PaymentGatewayInterface {
 
     }
 
-
     /**
-     * @param $parameters
      * @throws IndipayParametersMissingException
      */
     public function checkParameters($parameters)
@@ -120,8 +133,4 @@ class InstaMojoGateway implements PaymentGatewayInterface {
         }
 
     }
-
-
-
-
 }
