@@ -1,23 +1,32 @@
-<?php namespace Softon\Indipay\Gateways;
+<?php
 
+namespace Abounaja\Indipay\Gateways;
+
+use Abounaja\Indipay\Exceptions\IndipayParametersMissingException;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\View;
-use Softon\Indipay\Exceptions\IndipayParametersMissingException;
 
-class PayUMoneyGateway implements PaymentGatewayInterface {
+class PayUMoneyGateway implements PaymentGatewayInterface
+{
+    protected $parameters = [];
 
-    protected $parameters = array();
     protected $testMode = false;
+
     protected $merchantKey = '';
+
     protected $salt = '';
+
     protected $hash = '';
+
     protected $liveEndPoint = 'https://secure.payu.in/_payment';
+
     protected $testEndPoint = 'https://test.payu.in/_payment';
+
     public $response = '';
 
-    function __construct()
+    public function __construct()
     {
         $this->merchantKey = Config::get('indipay.payumoney.merchantKey');
         $this->salt = Config::get('indipay.payumoney.salt');
@@ -31,12 +40,12 @@ class PayUMoneyGateway implements PaymentGatewayInterface {
 
     public function getEndPoint()
     {
-        return $this->testMode?$this->testEndPoint:$this->liveEndPoint;
+        return $this->testMode ? $this->testEndPoint : $this->liveEndPoint;
     }
 
     public function request($parameters)
     {
-        $this->parameters = array_merge($this->parameters,$parameters);
+        $this->parameters = array_merge($this->parameters, $parameters);
 
         $this->checkParameters($this->parameters);
 
@@ -53,16 +62,16 @@ class PayUMoneyGateway implements PaymentGatewayInterface {
     {
 
         Log::info('Indipay Payment Request Initiated: ');
-        return View::make('indipay::payumoney')->with('hash',$this->hash)
-                             ->with('parameters',$this->parameters)
-                             ->with('endPoint',$this->getEndPoint());
+
+        return View::make('indipay::payumoney')->with('hash', $this->hash)
+            ->with('parameters', $this->parameters)
+            ->with('endPoint', $this->getEndPoint());
 
     }
 
-
     /**
      * Check Response
-     * @param $request
+     *
      * @return array
      */
     public function response($request)
@@ -71,16 +80,14 @@ class PayUMoneyGateway implements PaymentGatewayInterface {
 
         $response_hash = $this->decrypt($response);
 
-        if($response_hash!=$response['hash']){
+        if ($response_hash != $response['hash']) {
             return 'Hash Mismatch Error';
         }
 
         return $response;
     }
 
-
     /**
-     * @param $parameters
      * @throws IndipayParametersMissingException
      */
     public function checkParameters($parameters)
@@ -105,16 +112,15 @@ class PayUMoneyGateway implements PaymentGatewayInterface {
 
     /**
      * PayUMoney Encrypt Function
-     *
      */
     protected function encrypt()
     {
         $this->hash = '';
-        $hashSequence = "key|txnid|amount|productinfo|firstname|email|udf1|udf2|udf3|udf4|udf5|udf6|udf7|udf8|udf9|udf10";
+        $hashSequence = 'key|txnid|amount|productinfo|firstname|email|udf1|udf2|udf3|udf4|udf5|udf6|udf7|udf8|udf9|udf10';
         $hashVarsSeq = explode('|', $hashSequence);
         $hash_string = '';
 
-        foreach($hashVarsSeq as $hash_var) {
+        foreach ($hashVarsSeq as $hash_var) {
             $hash_string .= isset($this->parameters[$hash_var]) ? $this->parameters[$hash_var] : '';
             $hash_string .= '|';
         }
@@ -126,35 +132,29 @@ class PayUMoneyGateway implements PaymentGatewayInterface {
     /**
      * PayUMoney Decrypt Function
      *
-     * @param $plainText
-     * @param $key
+     * @param  $plainText
+     * @param  $key
      * @return string
      */
     protected function decrypt($response)
     {
 
-        $hashSequence = "status||||||udf5|udf4|udf3|udf2|udf1|email|firstname|productinfo|amount|txnid|key";
+        $hashSequence = 'status||||||udf5|udf4|udf3|udf2|udf1|email|firstname|productinfo|amount|txnid|key';
         $hashVarsSeq = explode('|', $hashSequence);
-        $hash_string = $this->salt."|";
+        $hash_string = $this->salt.'|';
 
-        foreach($hashVarsSeq as $hash_var) {
+        foreach ($hashVarsSeq as $hash_var) {
             $hash_string .= isset($response[$hash_var]) ? $response[$hash_var] : '';
             $hash_string .= '|';
         }
 
-        $hash_string = trim($hash_string,'|');
+        $hash_string = trim($hash_string, '|');
 
         return strtolower(hash('sha512', $hash_string));
     }
 
-
-
     public function generateTransactionID()
     {
-        return substr(hash('sha256', mt_rand() . microtime()), 0, 20);
+        return substr(hash('sha256', mt_rand().microtime()), 0, 20);
     }
-
-
-
-
 }
