@@ -1,7 +1,12 @@
-<?php namespace Softon\Indipay\Gateways;
+<?php
 
-interface PaymentGatewayInterface {
+namespace Abounaja\Indipay\Gateways;
+
+interface PaymentGatewayInterface
+{
     public function request($parameters);
+
     public function send();
+
     public function response($request);
 }
