@@ -1,51 +1,52 @@
-<?php namespace Softon\Indipay;
+<?php
+
+namespace Abounaja\Indipay;
 
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\ServiceProvider;
 
-class IndipayServiceProvider extends ServiceProvider {
+class IndipayServiceProvider extends ServiceProvider
+{
+    /**
+     * Indicates if loading of the provider is deferred.
+     *
+     * @var bool
+     */
+    protected $defer = false;
 
-	/**
-	 * Indicates if loading of the provider is deferred.
-	 *
-	 * @var bool
-	 */
-	protected $defer = false;
-
-	/**
-	 * Register the service provider.
-	 *
-	 * @return void
-	 */
-	public function register()
-	{
+    /**
+     * Register the service provider.
+     *
+     * @return void
+     */
+    public function register()
+    {
         $gateway = Config::get('indipay.gateway');
-        $this->app->bind('indipay', 'Softon\Indipay\Indipay');
+        $this->app->bind('indipay', 'Abounaja\Indipay\Indipay');
 
-        $this->app->bind('Softon\Indipay\Gateways\PaymentGatewayInterface','Softon\Indipay\Gateways\\'.$gateway.'Gateway');
-	}
+        $this->app->bind('Abounaja\Indipay\Gateways\PaymentGatewayInterface', 'Abounaja\Indipay\Gateways\\'.$gateway.'Gateway');
+    }
 
-
-    public function boot(){
+    public function boot()
+    {
         $this->publishes([
             __DIR__.'/config/config.php' => base_path('config/indipay.php'),
             __DIR__.'/views/middleware.blade.php' => base_path('app/Http/Middleware/VerifyCsrfMiddleware.php'),
         ]);
 
-		$this->loadViewsFrom(__DIR__.'/views', 'indipay');
+        $this->loadViewsFrom(__DIR__.'/views', 'indipay');
 
     }
 
-	/**
-	 * Get the services provided by the provider.
-	 *
-	 * @return array
-	 */
-	public function provides()
-	{
-		return [
+    /**
+     * Get the services provided by the provider.
+     *
+     * @return array
+     */
+    public function provides()
+    {
+        return [
 
         ];
-	}
-
+    }
 }
