@@ -11,11 +11,11 @@ return [
 
     'gateway' => 'Mocker',                // Replace with the name of default gateway you want to use
 
-    'testMode'  => true,                   // True for Testing the Gateway [For production false]
+    'testMode' => true,                   // True for Testing the Gateway [For production false]
 
     'ccavenue' => [                         // CCAvenue Parameters
-        'merchantId'  => env('INDIPAY_MERCHANT_ID', ''),
-        'accessCode'  => env('INDIPAY_ACCESS_CODE', ''),
+        'merchantId' => env('INDIPAY_MERCHANT_ID', ''),
+        'accessCode' => env('INDIPAY_ACCESS_CODE', ''),
         'workingKey' => env('INDIPAY_WORKING_KEY', ''),
 
         // Should be route address for url() function
@@ -27,8 +27,8 @@ return [
     ],
 
     'payumoney' => [                         // PayUMoney Parameters
-        'merchantKey'  => env('INDIPAY_MERCHANT_KEY', ''),
-        'salt'  => env('INDIPAY_SALT', ''),
+        'merchantKey' => env('INDIPAY_MERCHANT_KEY', ''),
+        'salt' => env('INDIPAY_SALT', ''),
         'workingKey' => env('INDIPAY_WORKING_KEY', ''),
 
         // Should be route address for url() function
@@ -37,7 +37,7 @@ return [
     ],
 
     'ebs' => [                         // EBS Parameters
-        'account_id'  => env('INDIPAY_MERCHANT_ID', ''),
+        'account_id' => env('INDIPAY_MERCHANT_ID', ''),
         'secretKey' => env('INDIPAY_WORKING_KEY', ''),
 
         // Should be route address for url() function
@@ -45,7 +45,7 @@ return [
     ],
 
     'citrus' => [                         // Citrus Parameters
-        'vanityUrl'  => env('INDIPAY_CITRUS_VANITY_URL', ''),
+        'vanityUrl' => env('INDIPAY_CITRUS_VANITY_URL', ''),
         'secretKey' => env('INDIPAY_WORKING_KEY', ''),
 
         // Should be route address for url() function
@@ -53,25 +53,25 @@ return [
         'notifyUrl' => env('INDIPAY_SUCCESS_URL', 'indipay/response'),
     ],
 
-    'instamojo' =>  [
-        'api_key' => env('INSTAMOJO_API_KEY',''),
-        'auth_token' => env('INSTAMOJO_AUTH_TOKEN',''),
+    'instamojo' => [
+        'api_key' => env('INSTAMOJO_API_KEY', ''),
+        'auth_token' => env('INSTAMOJO_AUTH_TOKEN', ''),
         'redirectUrl' => env('INDIPAY_REDIRECT_URL', 'indipay/response'),
     ],
 
-    'mocker' =>  [
-        'service' => env('MOCKER_SERVICE','default'),
+    'mocker' => [
+        'service' => env('MOCKER_SERVICE', 'default'),
         'redirect_url' => env('MOCKER_REDIRECT_URL', 'indipay/response'),
     ],
 
-    'zapakpay' =>  [
-        'merchantIdentifier' => env('ZAPAKPAY_MERCHANT_ID',''),
+    'zapakpay' => [
+        'merchantIdentifier' => env('ZAPAKPAY_MERCHANT_ID', ''),
         'secret' => env('ZAPAKPAY_SECRET', ''),
         'returnUrl' => env('ZAPAKPAY_RETURN_URL', 'indipay/response'),
     ],
 
-    'paytm' =>  [
-        'MERCHANT_KEY' => env('PAYTM_MERCHANT_KEY',''),
+    'paytm' => [
+        'MERCHANT_KEY' => env('PAYTM_MERCHANT_KEY', ''),
         'MID' => env('PAYTM_MID', ''),
         'CHANNEL_ID' => env('PAYTM_CHANNEL_ID', 'WEB'),
         'WEBSITE' => env('PAYTM_WEBSITE', 'WEBSTAGING'),
@@ -81,11 +81,7 @@ return [
 
     // Add your response link here. In Laravel 5.2+ you may use the VerifyCsrf Middleware.
     'remove_csrf_check' => [
-        'indipay/response'
+        'indipay/response',
     ],
-
-
-
-
 
 ];
